Conflicts:
	Makefile
	funs/all.hpp
	funs/exacttoinexact.cpp
	funs/imagpart.cpp
	funs/inexacttoexact.cpp
	funs/realpart.cpp
	parsers/identifier.cpp
	types/complex.cpp
	types/identifier.cpp
	utility/complextype.cpp
	utility/rationaltype.cpp

+++ types/float.hpp
@@ -3,6 +3,6 @@
 #include "arch.hpp"
 #include <string>
 
-PARSER_DECLARATION(FloatParser, Float, double)
+PARSER_DECLARATION(FloatParser, Float, long double)
     
 #endif
